PWR003: Simplify the Fortran example
We decided to use a separate file for the module because we thought that
moving the implementation to a different TU would allow hoisting of
`gravityOf` when marked with `pure`. This is not the case. `pure` has a
complete different meaning in Fortran. Let us therefore simplify the
example and use one source for the example and one source for the
solution.