PWR072: Suggest splitting the variable init
Instead of adding an explicit `save` attribute.

This update results from taking into account that, in most times, the
`save` behavior is unintentional and thus a source of bugs, instead of
something that programmers want to leverage in their codes.

Note that the benchmark code has not been updated. If a code is
incorrect due to the implicit `save` behavior, and the issue is resolved
by splitting the initialization, it does not make sense to compare the
performance of an incorrect code vs. a correct one. In contrast, there's
still value in benchmarking the particular scenario where the `save`
behavior is desired, comparing the performance of making it explicit vs.
keeping it implicit.