Rebrand the Open Catalog to include `correctness`
This commit introduces a new `correctness` check category, complementing
the existing `modernization` and `optimization` categories. The
motivation is to prevent grouping under `modernization` checks that are
not related to updating legacy code.

Take, for instance, PWR075. This check advises against relying on
compiler-specific extensions. These are not generally considered legacy
constructs, but they still pose a risk to code portability and consistent
behavior across different environments. Consequently, PWR075 is better
suited to the new `correctness` category.