Web: Sort `DataTable` columns with ticks
When the user clicks on a column, the `DataTable` applies an ascending
sorting. By default, this makes rows with no ticks appear before rows
with ticks. Switching this behavior is more natural as it allows, for
example, to click the `C` column and see first checks that are
applicable to the language (i.e., with a tick).