Replace `parameter` with `argument`
When referring to a procedure's arguments.

Various naming aspects of the Open Catalog aren't applicable to Fortran.
For example, multiple recommendations used `parameter` to refer to a
procedure's arguments; this is valid for C/C++, but confusing for
Fortran since `parameter` is a reserved keyword. Let's consistently use
`argument`, which is valid for all languages.

There are other pending issues, like the fact that C/C++ only have
`functions`, while Fortran has both `functions` and `subroutines`, which
can be referred to as `procedures`. However, this is not an usual term
for C/C++. We'll evaluate how to address these more conflicting issues
later on, as there's no simple solution for all languages.